pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Flask URL shortener"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "gunicorn>=22.0.0",
    "psycopg[binary]>=3.1.0",
    "python-dotenv>=1.0.0",
    "sqlmodel>=0.0.16",
]

[dependency-groups]
dev = [
    "pre-commit>=4.6.0",
    "pytest>=8.0.0",
    "ruff>=0.4.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
